fix tests, use real diff (slow) (#203022)
