+++ src/vs/workbench/contrib/inlineChat/test/browser/inlineChatController.test.ts
@@ -40,6 +40,8 @@ import { CTX_INLINE_CHAT_USER_DID_EDIT, EditMode, IInlineChatService, InlineChat
 import { InlineChatServiceImpl } from 'vs/workbench/contrib/inlineChat/common/inlineChatServiceImpl';
 import { workbenchInstantiationService } from 'vs/workbench/test/browser/workbenchTestServices';
 import { EditOperation } from 'vs/editor/common/core/editOperation';
+import { TestWorkerService } from './testWorkerService';
+import { IEditorWorkerService } from 'vs/editor/common/services/editorWorker';
 
 suite('InteractiveChatController', function () {
 	class TestController extends InlineChatController {
@@ -106,6 +108,7 @@ suite('InteractiveChatController', function () {
 		configurationService.setUserConfiguration('editor', {});
 
 		const serviceCollection = new ServiceCollection(
+			[IEditorWorkerService, new SyncDescriptor(TestWorkerService)],
 			[IContextKeyService, contextKeyService],
 			[IInlineChatService, inlineChatService],
 			[IDiffProviderFactoryService, new SyncDescriptor(TestDiffProviderFactoryService)],
@@ -263,7 +266,7 @@ suite('InteractiveChatController', function () {
 
 		const session = inlineChatSessionService.getSession(editor, editor.getModel()!.uri);
 		assert.ok(session);
-		assert.deepStrictEqual(session.wholeRange.value, new Range(1, 1, 1, 6));
+		assert.deepStrictEqual(session.wholeRange.value, new Range(1, 1, 1, 10 /* line length */));
 
 		editor.setSelection(new Range(2, 1, 2, 1));
 		editor.trigger('test', 'type', { text: 'a' });
@@ -299,19 +302,19 @@ suite('InteractiveChatController', function () {
 		store.add(d);
 		ctrl = instaService.createInstance(TestController, editor);
 		const p = ctrl.waitFor(TestController.INIT_SEQUENCE);
-		const r = ctrl.run({ message: 'Hello', autoSend: false });
+		const r = ctrl.run({ message: 'GENGEN', autoSend: false });
 
 		await p;
 
 		const session = inlineChatSessionService.getSession(editor, editor.getModel()!.uri);
 		assert.ok(session);
-		assert.deepStrictEqual(session.wholeRange.value, new Range(3, 1, 3, 3));
+		assert.deepStrictEqual(session.wholeRange.value, new Range(3, 1, 3, 3)); // initial
 
 		ctrl.acceptInput();
 
 		await ctrl.waitFor([State.MAKE_REQUEST, State.APPLY_RESPONSE, State.SHOW_RESPONSE, State.WAIT_FOR_INPUT]);
 
-		assert.deepStrictEqual(session.wholeRange.value, new Range(4, 1, 4, 3));
+		assert.deepStrictEqual(session.wholeRange.value, new Range(1, 1, 4, 3));
 
 		await ctrl.cancelSession();
 		await r;
